Refactoring: GridCreatingConditionTriangle::RemovePolygonVertexCommand implementation is separated from gridcreatingconditiontriangle.cpp.
