+++ libs/gridcreatingcondition/triangle/triangle.pro
@@ -116,7 +116,8 @@ HEADERS += gcc_triangle_global.h \
            private/gridcreatingconditiontriangle_movepolygoncommand.h \
            private/gridcreatingconditiontriangle_movepolygonvertexcommand.h \
            private/gridcreatingconditiontriangle_movepolylinecommand.h \
-           private/gridcreatingconditiontriangle_movepolylinevertexcommand.h
+           private/gridcreatingconditiontriangle_movepolylinevertexcommand.h \
+           private/gridcreatingconditiontriangle_removepolygonvertexcommand.h
 FORMS += gridcreatingconditiontrianglepolygoncoordinateseditdialog.ui \
          gridcreatingconditiontrianglepolylinecoordinateseditdialog.ui \
          gridcreatingconditiontriangleremeshdialog.ui \
@@ -143,7 +144,8 @@ SOURCES += gridcreatingconditioncreatortriangle.cpp \
            private/gridcreatingconditiontriangle_movepolygoncommand.cpp \
            private/gridcreatingconditiontriangle_movepolygonvertexcommand.cpp \
            private/gridcreatingconditiontriangle_movepolylinecommand.cpp \
-           private/gridcreatingconditiontriangle_movepolylinevertexcommand.cpp
+           private/gridcreatingconditiontriangle_movepolylinevertexcommand.cpp \
+           private/gridcreatingconditiontriangle_removepolygonvertexcommand.cpp
 TRANSLATIONS += languages/iricGccTriangle_es_ES.ts \
                 languages/iricGccTriangle_fr_FR.ts \
                 languages/iricGccTriangle_id_ID.ts \
